requirements.txt
tensorflow>=2.8.0
keras_applications>=1.0.8
kapre>=0.3.7
